This is a bit of an experiment.
Instead of a resource pool that requires clients to acquire and hand
back a resource as they need it, this pool maintains one resource per
thread in a ThreadLocal. Releasing the resource when a thread terminates
is a bit tricky and solved here by using a timer thread which checks
whether threads have terminated.